Сгруппировал мапперы по пакетам.
