+++ src/main/java/ru/academicians/myhelper/repository/mapper/subscription/SubscriptionRecordRowMapper.java
@@ -1,4 +1,4 @@
-package ru.academicians.myhelper.repository.mapper;
+package ru.academicians.myhelper.repository.mapper.subscription;
 
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
