+++ src/main/java/ru/academicians/myhelper/repository/mapper/user/UserEntityRowMapper.java
@@ -1,4 +1,4 @@
-package ru.academicians.myhelper.repository.mapper;
+package ru.academicians.myhelper.repository.mapper.user;
 
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
@@ -15,7 +15,7 @@ public class UserEntityRowMapper implements RowMapper<User> {
         result.setId(rs.getLong("id"));
         result.setFirstName(rs.getString("first_name"));
         result.setLastName(rs.getString("last_name"));
-        result.setMiddle_name(rs.getString("middle_name"));
+        result.setMiddleName(rs.getString("middle_name"));
         return result;
     }
 }
